chore: Reorganize files
