+++ packages/blockly/core/keyboard_nav/navigation_policies/comment_bar_button_navigation_policy.ts
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {CommentBarButton} from '../comments/comment_bar_button.js';
-import type {IFocusableNode} from '../interfaces/i_focusable_node.js';
-import type {INavigationPolicy} from '../interfaces/i_navigation_policy.js';
+import {CommentBarButton} from '../../comments/comment_bar_button.js';
+import type {IFocusableNode} from '../../interfaces/i_focusable_node.js';
+import type {INavigationPolicy} from '../../interfaces/i_navigation_policy.js';
 
 /**
  * Set of rules controlling keyboard navigation from a CommentBarButton.
