+++ packages/blockly/core/workspace_svg.ts
@@ -60,9 +60,9 @@ import {hasBubble} from './interfaces/i_has_bubble.js';
 import type {IMetricsManager} from './interfaces/i_metrics_manager.js';
 import type {IToolbox} from './interfaces/i_toolbox.js';
 import {KeyboardMover} from './keyboard_nav/keyboard_mover.js';
+import {Navigator} from './keyboard_nav/navigators/navigator.js';
 import {LayerManager} from './layer_manager.js';
 import {Msg} from './msg.js';
-import {Navigator} from './navigator.js';
 import {Options} from './options.js';
 import * as Procedures from './procedures.js';
 import * as registry from './registry.js';
